% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{veglib_IGBP}
\alias{veglib_IGBP}
\title{IGBP vegetation library for VIC model}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 17 rows and 58 columns.
}
\usage{
veglib_IGBP
}
\description{
This library is originated from Python package \code{hydrate}.
}
\references{
\enumerate{
\item https://github.com/KMarkert/hydrate/blob/master/hydrate/lookups/veg_params.py
\item https://github.com/KMarkert/hydrate
}
}
\keyword{datasets}
