% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIC5-package.R
\docType{package}
\name{VIC5}
\alias{VIC5}
\alias{VIC5-package}
\title{VIC5: The Variable Infiltration Capacity (VIC) Model version 5 in R}
\description{
The Variable Infiltration Capacity (VIC) model is a macroscale hydrologic model that solves full water and energy balances, originally developed by Xu Liang at the University of Washington (UW). The version of VIC source code used is of 5.0.1 on <https://github.com/UW-Hydro/VIC/>, see Hamman et al. (2018). Development and maintenance of the current official version of the VIC model at present is led by the UW Hydro (Computational Hydrology group) in the Department of Civil and Environmental Engineering at UW. VIC is a research model and in its various forms it has been applied to most of the major river basins around the world, as well as globally <http://vic.readthedocs.io/en/master/Documentation/References/>. References: "Liang, X., D. P. Lettenmaier, E. F. Wood, and S. J. Burges (1994), A simple hydrologically based model of land surface water and energy fluxes for general circulation models, J. Geophys. Res., 99(D7), 14415-14428, <doi:10.1029/94JD00483>"; "Hamman, J. J., Nijssen, B., Bohn, T. J., Gergel, D. R., and Mao, Y. (2018), The Variable Infiltration Capacity model version 5 (VIC-5): infrastructure improvements for new applications and reproducibility, Geosci. Model Dev., 11, 3481-3496, <doi:10.5194/gmd-11-3481-2018>".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rpkgs/VIC5}
}

}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com} (\href{https://orcid.org/0000-0003-1836-8172}{ORCID})

Authors:
\itemize{
  \item Ruida Zhong \email{zrd2017@163.com} [contributor]
  \item Xiaohong Chen \email{eescxh@mail.sysu.edu.cn} [contributor]
  \item Zhaoli Wang \email{wangzhl@scut.edu.cn} [contributor]
  \item Chengguang Lai \email{laichg@scut.edu.cn} [contributor]
}

Other contributors:
\itemize{
  \item Joseph J Hamman (Contributor and maintainer of VIC source code) [contributor]
  \item Keith Cherkauer [contributor]
}

}
\keyword{internal}
